refactor(frontend): group admin API files under api/admin/
Flat api/admin*.ts (3 files + existing ai.ts) -> api/admin/ subfolder
with cleaner names (index, ai, worldViewImport, wikivoyageExtract).
Updates 17 import sites plus 3 intra-folder relative imports.

- api/admin.ts       -> api/admin/index.ts
- api/adminWikivoyageExtract.ts -> api/admin/wikivoyageExtract.ts
- api/adminWorldViewImport.ts   -> api/admin/worldViewImport.ts
- api/admin/ai.ts stays (already moved in PR-40)
